refactor(settings): split AppSettings into focused sub-components
- Extract AppSettingsTheme, AppSettingsNotifications, AppSettingsNavCard
- Move notification/push logic into useNotificationSettings composable
- AppSettings becomes thin orchestrator with lazy-loaded sub-components
